requirements.txt
django==1.10.6
djangorestframework==3.5.4
python-dateutil==2.6.0
